refactor: simplify truncation calculation in multiple subroutines for improved readability

+++ src/skyborn/spharm/src/multsmoothfact.f90
@@ -17,7 +17,7 @@ subroutine multsmoothfact(dataspec, dataspec_smooth, smooth, nlat, nmdim, nt)
     real :: smooth_factor
 
     ! Calculate truncation from nmdim
-    ntrunc = int(-1.5 + 0.5 * sqrt(9.0 - 8.0 * (1.0 - real(nmdim))))
+    ntrunc = -1.5 + 0.5 * sqrt(9.0 - 8.0 * (1.0 - real(nmdim)))
 
     ! Main computation loops - optimized for vectorization and cache efficiency
     !$OMP PARALLEL DO PRIVATE(nmstrt, m, n, nm, smooth_factor) &
